Dynamic Memory Allocator Learning Path
This learning path is for folks who know some C programming and
have maybe used malloc and free before . It gives them a tour of
a simple dynamic memory allocator and discusses aspects of how it works
and how it could be improved.

The overall goal is to demystify the subject so there is a
lower barrier to entry if folks want to get onto more advanced
memory allocation topics.

For exampe, in future I would like to build on this by adding
Arm Memory Tagging to the allocator in a separate learning path.

The concepts are universal but the build instructions are for
a Linux system, so I've put it in the laptops-and-desktops
category.