Fix UnboundLocalError for username when testing multiple learning path pages
Prior to this change, when I tested a learning path that had >1 page using
`ubuntu:latest` I got:
```
Traceback (most recent call last):
  File "./tools/maintenance.py", line 171, in <module>
    main()
  File "./tools/maintenance.py", line 148, in main
    check_lp(args.instructions, args.link, args.debug)
  File "./tools/maintenance.py", line 60, in check_lp
    res.append(check.check(lp_path + "/" + i[0], start=launch, stop=terminate))
  File "/home/davspi01/work/open_source/arm-learning-paths/tools/check.py", line 215, in check
    cmd = ["docker cp {} test_{}:/home/{}/".format(fn, k, username)]
UnboundLocalError: local variable 'username' referenced before assignment
```

This happened because username is only set when `start` is true and this
only happens for the first page.

```
[INFO]	Checking how-to-1.md
> /home/davspi01/work/open_source/arm-learning-paths/tools/check.py(83)check()
-> with open(json_file) as jf:
(Pdb) print(start, stop)
True False
(Pdb) c
[INFO]	Checking how-to-3.md
> /home/davspi01/work/open_source/arm-learning-paths/tools/check.py(83)check()
-> with open(json_file) as jf:
(Pdb) print(start, stop)
False False
(Pdb) c
[INFO]	Checking how-to-2.md
> /home/davspi01/work/open_source/arm-learning-paths/tools/check.py(83)check()
-> with open(json_file) as jf:
(Pdb) print(start, stop)
False True
(Pdb) c
```

Where it was, `username` didn't really need to be a variable because it was
hardcoded into the docker commands for each branch of the if.

So I've moved the username lookup later where we check the image we're going
to run on (which happens regardless of whether we start a new image).