Update _next-steps.md
next_step_guidance: You should now be able to test the `restrict` keyword on your own or other open-source code and discover potential optimizations! - this sentence should be added to review.md as per comments already made in that commit. Next step guidance should be if there are other learning paths or documentation in, say, developer.arm.com that's useful to add to the newly acquired language.

recommended_path: /learning-paths/embedded-systems/ - can we be more specific here. Is there a particular learning path that's helpful? Please can this be amended.

further_reading:
    - resource:
        title: Wikipedia restrict entry
        link: https://en.wikipedia.org/wiki/Restrict
        type: documentation
We do not recommend adding links to Wikipedia. It's too general and doesn't add to the learning path. 

    - resource:
        title: Godbolt restrict tests - is this more useful as a next_step_guidance? What do you want the learner to be able to do with these tests?