[hugo-server] Usability improvement.
As the hugo-server.sh script is to be used by all LP authors, working
on drafts, simplify the handling of the draft by forcing hugo to build
the site including the drafts. Publishing is only dealt with by a few
people who can do what needs to be done with drafts.