[Workflows] Only run the `sweep` job on the main repository.
My understanding is that this job collectes statistics for our own
usage. There is no benefit to the contributors working on clones when
this job fails (which it does at the moment).

This commit enables the `last-reviewed-cron.yml` workflow to only run
when the repository is our main LP repository.